A query may be specified to be run before map reduce jobs.
The options passed into the job can now contain a "prefilter". This
should be specified as a Hash that can be executed as a MongoDB query.
This query will be run to filter the records that the map reduce job
will run on.

It should be noted that after this change, Cypress will need to pass
in the test_id as a prefilter. That will no longer be handled
separately.