Fixes the creation of the BSON::ObjectId
The latest version of Moped moves away from its own implementation of
BSON and switches back to the default. With this change, there are
some API changes. This is a spot where we are trying to create a BSON
ObjectId through the old API, where we need to use from_string.