Now assigning providers that overlap with the reporting period
Previously, a provider would be assigned to a patient cache result
only if it completely overlapped the measure period or overlapped the
end of the measure period. Providers that overlapped the start but not
the end of the reporting were not added to the provider performances
in the patient cache entry.

Now, any provider that has any overlap with the period of performance
will be applied to the patient cache result.