Adding the records collection to the set up for these tests
test_rollup_buffering will run a map reduce job. While the results of
the job do not matter to this particular test, it will fail if the
records collection is not there yet. My belief is that this test
passed previously because it was dependent on tests that ran before
it. Upgrading to minitest 5.4 probably highlighted this.