Adds a class to inject the effective start date
This allows us to use the effective start date to set the low value of
the measure period in every measure. This is used to determine which
patients are used within each measure.

If no effective start date is provided the measures use the original
code which calculates the low date based on the effective date + 60
seconds - 1 year.